Move widget to a FitParamView model
